combine appconfig
